finish Shr impls
