+++ rust/src/combinators/chain/mod.rs
@@ -1,11 +1,11 @@
 #[cfg(feature = "ffi")]
 mod ffi;
 
-use std::ops::Shr;
+mod shr;
 
 use crate::core::{
-    Domain, Function, Measure, Measurement, Metric, MetricSpace, PartialTransformation, PrivacyMap,
-    StabilityMap, Transformation,
+    Domain, Function, Measure, Measurement, Metric, MetricSpace, PrivacyMap, StabilityMap,
+    Transformation,
 };
 use crate::error::{Error, ErrorVariant, Fallible};
 use std::fmt::Debug;
@@ -316,234 +316,5 @@ mod tests {
         assert_eq!(d_out, 99);
 
         Ok(())
-    }
-}
-
-impl<DI, DX, TO, MI, MX, MO> Shr<Measurement<DX, TO, MX, MO>> for Transformation<DI, DX, MI, MX>
-where
-    DI: 'static + Domain,
-    DX: 'static + Domain,
-    TO: 'static,
-    MI: 'static + Metric,
-    MX: 'static + Metric,
-    MO: 'static + Measure,
-    (DI, MI): MetricSpace,
-    (DX, MX): MetricSpace,
-{
-    type Output = Fallible<Measurement<DI, TO, MI, MO>>;
-
-    fn shr(self, rhs: Measurement<DX, TO, MX, MO>) -> Self::Output {
-        make_chain_mt(&rhs, &self)
-    }
-}
-
-impl<DI, DX, TO, MI, MX, MO> Shr<Measurement<DX, TO, MX, MO>>
-    for Fallible<Transformation<DI, DX, MI, MX>>
-where
-    DI: 'static + Domain,
-    DX: 'static + Domain,
-    TO: 'static,
-    MI: 'static + Metric,
-    MX: 'static + Metric,
-    MO: 'static + Measure,
-    (DI, MI): MetricSpace,
-    (DX, MX): MetricSpace,
-{
-    type Output = Fallible<Measurement<DI, TO, MI, MO>>;
-
-    fn shr(self, rhs: Measurement<DX, TO, MX, MO>) -> Self::Output {
-        make_chain_mt(&rhs, &self?)
-    }
-}
-
-impl<DI, DX, DO, MI, MX, MO> Shr<Transformation<DX, DO, MX, MO>> for Transformation<DI, DX, MI, MX>
-where
-    DI: 'static + Domain,
-    DX: 'static + Domain,
-    DO: 'static + Domain,
-    MI: 'static + Metric,
-    MX: 'static + Metric,
-    MO: 'static + Metric,
-    (DI, MI): MetricSpace,
-    (DX, MX): MetricSpace,
-    (DO, MO): MetricSpace,
-{
-    type Output = Fallible<Transformation<DI, DO, MI, MO>>;
-
-    fn shr(self, rhs: Transformation<DX, DO, MX, MO>) -> Self::Output {
-        make_chain_tt(&rhs, &self)
-    }
-}
-
-impl<DI, DX, DO, MI, MX, MO> Shr<PartialTransformation<DX, DO, MX, MO>>
-    for Transformation<DI, DX, MI, MX>
-where
-    DI: 'static + Domain,
-    DX: 'static + Domain,
-    DO: 'static + Domain,
-    MI: 'static + Metric,
-    MX: 'static + Metric,
-    MO: 'static + Metric,
-    (DI, MI): MetricSpace,
-    (DX, MX): MetricSpace,
-    (DO, MO): MetricSpace,
-{
-    type Output = Fallible<Transformation<DI, DO, MI, MO>>;
-
-    fn shr(self, rhs: PartialTransformation<DX, DO, MX, MO>) -> Self::Output {
-        let rhs = (rhs).fix(self.output_domain.clone(), self.output_metric.clone())?;
-        make_chain_tt(&rhs, &self)
-    }
-}
-
-impl<DI, DO, MI, MO> Shr<PartialTransformation<DI, DO, MI, MO>> for (DI, MI)
-where
-    DI: 'static + Domain,
-    DO: 'static + Domain,
-    MI: 'static + Metric,
-    MO: 'static + Metric,
-    (DI, MI): MetricSpace,
-    (DO, MO): MetricSpace,
-{
-    type Output = Fallible<Transformation<DI, DO, MI, MO>>;
-
-    fn shr(self, rhs: PartialTransformation<DI, DO, MI, MO>) -> Self::Output {
-        (rhs).fix(self.0, self.1)
-    }
-}
-
-impl<DI, DX, DO, MI, MX, MO> Shr<PartialTransformation<DX, DO, MX, MO>>
-    for Fallible<Transformation<DI, DX, MI, MX>>
-where
-    DI: 'static + Domain,
-    DX: 'static + Domain,
-    DO: 'static + Domain,
-    MI: 'static + Metric,
-    MX: 'static + Metric,
-    MO: 'static + Metric,
-    (DI, MI): MetricSpace,
-    (DX, MX): MetricSpace,
-    (DO, MO): MetricSpace,
-{
-    type Output = Fallible<Transformation<DI, DO, MI, MO>>;
-
-    fn shr(self, rhs: PartialTransformation<DX, DO, MX, MO>) -> Self::Output {
-        let lhs = self?;
-        let rhs = (rhs).fix(lhs.output_domain.clone(), lhs.output_metric.clone())?;
-        make_chain_tt(&rhs, &lhs)
-    }
-}
-
-impl<DI, DX, DO, MI, MX, MO> Shr<Transformation<DX, DO, MX, MO>>
-    for Fallible<Transformation<DI, DX, MI, MX>>
-where
-    DI: 'static + Domain,
-    DX: 'static + Domain,
-    DO: 'static + Domain,
-    MI: 'static + Metric,
-    MX: 'static + Metric,
-    MO: 'static + Metric,
-    (DI, MI): MetricSpace,
-    (DX, MX): MetricSpace,
-    (DO, MO): MetricSpace,
-{
-    type Output = Fallible<Transformation<DI, DO, MI, MO>>;
-
-    fn shr(self, rhs: Transformation<DX, DO, MX, MO>) -> Self::Output {
-        make_chain_tt(&rhs, &self?)
-    }
-}
-
-impl<DI, TX, TO, MI, MO> Shr<Function<TX, TO>> for Measurement<DI, TX, MI, MO>
-where
-    DI: 'static + Domain,
-    TX: 'static,
-    TO: 'static,
-    MI: 'static + Metric,
-    MO: 'static + Measure,
-    (DI, MI): MetricSpace,
-{
-    type Output = Fallible<Measurement<DI, TO, MI, MO>>;
-
-    fn shr(self, rhs: Function<TX, TO>) -> Self::Output {
-        make_chain_pm(&rhs, &self)
-    }
-}
-
-impl<DI, TX, TO, MI, MO> Shr<Function<TX, TO>> for Fallible<Measurement<DI, TX, MI, MO>>
-where
-    DI: 'static + Domain,
-    TX: 'static,
-    TO: 'static,
-    MI: 'static + Metric,
-    MO: 'static + Measure,
-    (DI, MI): MetricSpace,
-{
-    type Output = Fallible<Measurement<DI, TO, MI, MO>>;
-
-    fn shr(self, rhs: Function<TX, TO>) -> Self::Output {
-        make_chain_pm(&rhs, &self?)
-    }
-}
-
-impl<DI, DX, DO, MI, MO, MTI, MTO> Shr<Transformation<DX, DO, MTI, MTO>>
-    for Measurement<DI, DX::Carrier, MI, MO>
-where
-    DI: 'static + Domain,
-    DX: 'static + Domain,
-    DO: 'static + Domain,
-    MI: 'static + Metric,
-    MO: 'static + Measure,
-    MTI: 'static + Metric,
-    MTO: 'static + Metric,
-    (DI, MI): MetricSpace,
-    (DX, MTI): MetricSpace,
-    (DO, MTO): MetricSpace,
-{
-    type Output = Fallible<Measurement<DI, DO::Carrier, MI, MO>>;
-
-    fn shr(self, rhs: Transformation<DX, DO, MTI, MTO>) -> Self::Output {
-        make_chain_pm(&rhs.function, &self)
-    }
-}
-
-impl<DI, DX, DO, MI, MO, MTI, MTO> Shr<Transformation<DX, DO, MTI, MTO>>
-    for Fallible<Measurement<DI, DX::Carrier, MI, MO>>
-where
-    DI: 'static + Domain,
-    DX: 'static + Domain,
-    DO: 'static + Domain,
-    MI: 'static + Metric,
-    MO: 'static + Measure,
-    MTI: 'static + Metric,
-    MTO: 'static + Metric,
-    (DI, MI): MetricSpace,
-    (DX, MTI): MetricSpace,
-    (DO, MTO): MetricSpace,
-{
-    type Output = Fallible<Measurement<DI, DO::Carrier, MI, MO>>;
-
-    fn shr(self, rhs: Transformation<DX, DO, MTI, MTO>) -> Self::Output {
-        make_chain_pm(&rhs.function, &self?)
-    }
-}
-
-#[cfg(test)]
-mod tests_shr {
-    use crate::measurements::make_base_discrete_laplace;
-    use crate::transformations::{
-        make_bounded_sum, make_cast_default, make_clamp, make_split_lines,
-    };
-
-    use super::*;
-
-    #[test]
-    fn test_shr() -> Fallible<()> {
-        (make_split_lines()?
-            >> make_cast_default()?
-            >> make_clamp((0, 1))?
-            >> make_bounded_sum((0, 1))?
-            >> make_base_discrete_laplace(1.)?)
-        .map(|_| ())
     }
 }
