+++ rust/src/core/mod.rs
@@ -306,6 +306,26 @@ impl<DI: Domain, TO, MI: Metric, MO: Measure> Measurement<DI, TO, MI, MO> {
     }
 }
 
+pub struct PartialMeasurement<DI: Domain, TO, MI: Metric, MO: Measure>(
+    Box<dyn FnOnce(DI, MI) -> Fallible<Measurement<DI, TO, MI, MO>>>,
+)
+where
+    (DI, MI): MetricSpace;
+
+impl<DI: Domain, TO, MI: Metric, MO: Measure> PartialMeasurement<DI, TO, MI, MO>
+where
+    (DI, MI): MetricSpace,
+{
+    pub fn new(
+        partial: impl FnOnce(DI, MI) -> Fallible<Measurement<DI, TO, MI, MO>> + 'static,
+    ) -> Self {
+        Self(Box::new(partial))
+    }
+    pub fn fix(self, input_domain: DI, input_metric: MI) -> Fallible<Measurement<DI, TO, MI, MO>> {
+        (self.0)(input_domain, input_metric)
+    }
+}
+
 pub trait MetricSpace {
     fn check(&self) -> bool;
     fn assert_compatible(&self) -> Fallible<()> {
@@ -336,32 +356,6 @@ pub struct Transformation<DI: Domain, DO: Domain, MI: Metric, MO: Metric> {
     pub input_metric: MI,
     pub output_metric: MO,
     pub stability_map: StabilityMap<MI, MO>,
-}
-
-pub struct PartialTransformation<DI: Domain, DO: Domain, MI: Metric, MO: Metric>(
-    Box<dyn FnOnce(DI, MI) -> Fallible<Transformation<DI, DO, MI, MO>>>,
-)
-where
-    (DI, MI): MetricSpace,
-    (DO, MO): MetricSpace;
-
-impl<DI: Domain, DO: Domain, MI: Metric, MO: Metric> PartialTransformation<DI, DO, MI, MO>
-where
-    (DI, MI): MetricSpace,
-    (DO, MO): MetricSpace,
-{
-    pub fn new(
-        partial: impl FnOnce(DI, MI) -> Fallible<Transformation<DI, DO, MI, MO>> + 'static,
-    ) -> Self {
-        Self(Box::new(partial))
-    }
-    pub fn fix(
-        self,
-        input_domain: DI,
-        input_metric: MI,
-    ) -> Fallible<Transformation<DI, DO, MI, MO>> {
-        (self.0)(input_domain, input_metric)
-    }
 }
 
 impl<DI: Domain, DO: Domain, MI: Metric, MO: Metric> Transformation<DI, DO, MI, MO>
@@ -424,6 +418,32 @@ impl<DI: Domain, DO: Domain, MI: Metric, MO: Metric> Transformation<DI, DO, MI, 
         MO::Distance: TotalOrd,
     {
         d_out.total_ge(&self.map(d_in)?)
+    }
+}
+
+pub struct PartialTransformation<DI: Domain, DO: Domain, MI: Metric, MO: Metric>(
+    Box<dyn FnOnce(DI, MI) -> Fallible<Transformation<DI, DO, MI, MO>>>,
+)
+where
+    (DI, MI): MetricSpace,
+    (DO, MO): MetricSpace;
+
+impl<DI: Domain, DO: Domain, MI: Metric, MO: Metric> PartialTransformation<DI, DO, MI, MO>
+where
+    (DI, MI): MetricSpace,
+    (DO, MO): MetricSpace,
+{
+    pub fn new(
+        partial: impl FnOnce(DI, MI) -> Fallible<Transformation<DI, DO, MI, MO>> + 'static,
+    ) -> Self {
+        Self(Box::new(partial))
+    }
+    pub fn fix(
+        self,
+        input_domain: DI,
+        input_metric: MI,
+    ) -> Fallible<Transformation<DI, DO, MI, MO>> {
+        (self.0)(input_domain, input_metric)
     }
 }
